Implement dry run mode.

+++ src/index.ts
@@ -59,7 +59,11 @@ export async function generate(opts: GenerateOpts): Promise<void> {
 		});
 	}
 
-	const file = path.join(path.dirname(configFile), config.outfile);
-	debug('writing schema def to file', file);
-	await fs.promises.writeFile(file, result, 'utf-8');
+	if (!opts.dryRun) {
+		const file = path.join(path.dirname(configFile), config.outfile);
+		await fs.promises.writeFile(file, result, 'utf-8');
+		console.log('wrote schema to file', file);
+	} else {
+		console.log('// dry run mode\n', result);
+	}
 }
